+++ include/NeoFOAM/NeoFOAM.hpp
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: MIT
 // SPDX-FileCopyrightText: 2023 NeoFOAM authors
 
-#include "core/Error.hpp"
-#include "core/Info.hpp"
-#include "core/Time.hpp"
+#include "core.hpp"
+
+#include "finiteVolume/cellCentred.hpp"