+++ include/NeoFOAM/core/primitives/tensor.hpp
@@ -2,10 +2,4 @@
 // SPDX-FileCopyrightText: 2023 NeoFOAM authors
 #pragma once
 
-#include <scalar.hpp>
-
-class tensor
-{
-public:
-private:
-};
+#include "scalar.hpp"